test cases refactored with no default pick of QA mongo

+++ src/test/java/analytics/bolt/PersistDCBoltTest.java
@@ -45,13 +45,13 @@ public class PersistDCBoltTest {
 		date = simpleDateFormat.format(new Date());
 	}
 
-	@After
+/*	@After
 	public void tearDown() throws Exception {
 		if(db.toString()=="FongoDB.test")
 			   db.dropDatabase();
 			  else
 			   Assert.fail("Something went wrong. Tests connected to " + db.toString());
-	}
+	}*/
 
 	@Test
 	public void testPersistWhenMemberNotExist() throws JSONException {
@@ -228,10 +228,13 @@ public class PersistDCBoltTest {
 	}
 	
 	
-	
 	@AfterClass
-	public static void cleanUp(){
-		db.dropDatabase();
+	public static void teardown() {
+		if(db.toString().equalsIgnoreCase("FongoDB.test"))
+			   db.dropDatabase();
+			  else
+			   Assert.fail("Something went wrong. Tests connected to " + db.toString());
 	}
+
 
 }
